# openapi.yaml
openapi: 3.0.0
info:
  title: My Collection
  description: >-
    ### Welcome to Postman! This is your first collection.


    Collections are your starting point for building and testing APIs. You can
    use this one to:


    • Group related requests  

    • Test your API in real-world scenarios  

    • Document and share your requests


    Update the name and overview whenever you’re ready to make it yours.


    [Learn more about Postman
    Collections.](https://learning.postman.com/docs/collections/collections-overview/)
  version: 1.0.0
servers:
  - url: http://{{baseurl}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /api/users/reports:
    post:
      tags:
        - CSC309 Project > Users > Reports
      summary: Create report
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                reason: Test report, the post with id 1 is tested again.
                targetType: THREAD
                targetId: 1
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        undefined:
          content:
            text/plain:
              schema:
                type: string
              examples:
                example-0:
                  summary: Report thread 1
                  value: ''
                example-1:
                  summary: Report thread 1
                  value: ''
    get:
      tags:
        - CSC309 Project > Users > Reports
      summary: 5 reports on post 1
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          example: '1'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /api/users/follow:
    post:
      tags:
        - CSC309 Project > Users > Follow
      summary: Follow a user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                toFollowId: 1
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /api/users/feed:
    get:
      tags:
        - CSC309 Project > Users > Feed
      summary: Get feed
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          example: '1'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/users/signup:
    post:
      tags:
        - CSC309 Project > Users > Create users
      summary: Create second user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                email: uoft@userexample.com
                username: Eric
                password: '1234'
                role: USER
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /api/users/login:
    post:
      tags:
        - CSC309 Project > Users > Ban Users
      summary: User login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                username: Jaic
                password: '1234'
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...